pyproject.toml
[project]
name = "kaarel-bet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.109.1",
    "requests>=2.31.0",
    "names-dataset>=3.0.0",
    "pyyaml>=6.0.0",
    "dotenv>=0.9.9",
]

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
    "pyright[nodejs]>=1.1.0",
    "pre-commit>=4.3.0",
]

[tool.poe.tasks]
# run with eg `uv run poe fmt`
fmt = "ruff format"
lint = "ruff check --fix"
typecheck = "pyright"
test = "pytest"
# run all the above
all = [ {ref="fmt"}, {ref="lint"}, {ref="typecheck"}, {ref="test"} ]

[tool.pyright]
# Basic pyright configuration
include = ["kaarel-bet"]
exclude = ["**/__pycache__", "**/.venv"]
pythonVersion = "3.13"
typeCheckingMode = "basic"


